Use single-precision float-point values instead of double-precision values whenever data is being handled that will be written into the property value of a config element.  The C++ code is expecting single-precision values, and rounding errors tend to be exacerbated when using double-precision values.  Some double-precision values are still used for internal calcuations.
git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@15387 08b38cba-cd3b-11de-854e-f91c5b6e4272

+++ modules/vrjuggler/vrjconfig/customeditors/display_window/org/vrjuggler/vrjconfig/customeditors/display_window/DisplayWindowFrame.java
@@ -550,29 +550,29 @@ public class DisplayWindowFrame
 
             Point3D[] corners = dlg.getCorners();
             mSelectedViewport.setProperty("lower_left_corner", 0,
-                                          new Double(corners[0].x));
+                                          new Float(corners[0].x));
             mSelectedViewport.setProperty("lower_left_corner", 1,
-                                          new Double(corners[0].y));
+                                          new Float(corners[0].y));
             mSelectedViewport.setProperty("lower_left_corner", 2,
-                                          new Double(corners[0].z));
+                                          new Float(corners[0].z));
             mSelectedViewport.setProperty("lower_right_corner", 0,
-                                          new Double(corners[1].x));
+                                          new Float(corners[1].x));
             mSelectedViewport.setProperty("lower_right_corner", 1,
-                                          new Double(corners[1].y));
+                                          new Float(corners[1].y));
             mSelectedViewport.setProperty("lower_right_corner", 2,
-                                          new Double(corners[1].z));
+                                          new Float(corners[1].z));
             mSelectedViewport.setProperty("upper_right_corner", 0,
-                                          new Double(corners[2].x));
+                                          new Float(corners[2].x));
             mSelectedViewport.setProperty("upper_right_corner", 1,
-                                          new Double(corners[2].y));
+                                          new Float(corners[2].y));
             mSelectedViewport.setProperty("upper_right_corner", 2,
-                                          new Double(corners[2].z));
+                                          new Float(corners[2].z));
             mSelectedViewport.setProperty("upper_left_corner", 0,
-                                          new Double(corners[3].x));
+                                          new Float(corners[3].x));
             mSelectedViewport.setProperty("upper_left_corner", 1,
-                                          new Double(corners[3].y));
+                                          new Float(corners[3].y));
             mSelectedViewport.setProperty("upper_left_corner", 2,
-                                          new Double(corners[3].z));
+                                          new Float(corners[3].z));
 
             mSelectedViewport.setProperty("tracked", 0, dlg.isTracked());
